Allow identical requests by tracking order of occurrence of requests
Right now PHP-VCR only tracks one occurrence of a given request
per cassette. An example of why this is troublesome is given below.

* (1) Request a resource from an external endpoint (GET /resource)
* (2) Request an update to that resource (PUT /resource)
* (3) Re-request the resource to confirm it was persisted remotely (GET /resource)

Requests (1) and (3) are identical, however the responses are expected
to be different. PHP-VCR will only record (1) and use the saved version
on (3).

This commit keeps track of the order of occurrence of requests, meaning
(1) is the 1st occurrence of (GET /resource) and (3) is the 2nd. When
the test is re-run, when running (3) PHP-VCR will now look for a
saved request which is the 2nd (GET /resource).

See #132