enabled tests to simulate new iteration of test cases
tests were failing because between recording responses and testing
playback the index table was not reset. Since introducing the index
table we keep recording responses. Next time the test case is run,
the index will have reset and then the cassette will look for recordings
instead of recording new entries. This commit will simulate the
"end" of test cases and the "closing" of a cassette.