Fix curl reset (#248)
### Context

When using PHP-VCR with Guzzle 6 with async requests, the first request is playing correctly but the next request is putting Guzzle in an infinite loop.

It took me a whole day but I traced back this problem to the way the "curl_reset" hook is written. It clears the request but not the response. Hence, when another request is performed (even on a different URL) the same response is sent. This weird behaviour is causing an infinite loop in Guzzle 6.

This problem is probably exactly the same as the issue #211. This PR should fix it.

### What has been done

The first commit of this PR contains only the failing test (to showcase the problem). You see that the test never completes (because of the infinite loop): https://travis-ci.org/php-vcr/php-vcr/jobs/366140696

The second commit is the fix. I simply added one line in the curl_reset hook to remove the response as well as the request.

Note: this PR is based on the branch from PR #246 since I need unit tests from Guzzle 6 that are not yet merged.