Fixed AbstractCodeTransform registration.
AbstractCodeTransform was actually misbehaving if several instances of the same XxxCodeTransform were to be created.
Indeed, a wrapper can be registered, but not unregistered. If 2 instances of CurlCodeTransform were to be created, the registration of the wrapper would be attempted twice. The second one would fail and try again but the wrapper is already registered. This PR registers only if the stream is indeed not already registered.