FIX: Videorecorder always raised index for identical requests, Cassette have to handle legacy entries
Because the Videorecorder increments the index for each identical requests it also passes it to the cassette,
so the cassette have to handle legacy stored entries. So if no index is present while loading recorded requests
we have to set the index to searched one (instead of zero) to always match (disabled the feature), when all
other matchers match.